Add tests for precomputed layouts
