+++ tests/pixeldataset/test_datastores.py
@@ -3,22 +3,26 @@
 Copyright © 2024 Pixelgen Technologies AB.
 """
 
+from pathlib import Path
 from unittest.mock import patch
+from zipfile import ZipFile
+
+import pandas as pd
+import polars as pl
+import pytest
 from anndata import AnnData
 from pandas.core.frame import DataFrame
-from pathlib import Path
-
-import pytest
 from pandas.testing import assert_frame_equal
 from pixelator.pixeldataset import PixelDataset
-
 from pixelator.pixeldataset.datastores import (
     FileFormatNotRecognizedError,
     PixelDataStore,
     ZipBasedPixelFile,
     ZipBasedPixelFileWithCSV,
     ZipBasedPixelFileWithParquet,
+    CannotOverwriteError,
 )
+from pixelator.pixeldataset.precomputed_layouts import PreComputedLayouts
 
 
 class TestPixelDataStore:
@@ -77,12 +81,8 @@ class TestPixelDataStore:
         result = dataset.file_format_version()
         assert result == 1
 
-    @pytest.mark.parametrize(
-        "datastore", [ZipBasedPixelFileWithParquet, ZipBasedPixelFileWithCSV]
-    )
-    def test_pixelfile_datastore_can_switch_between_reading_and_writing(
+    def test_pixelfile_datastore_trying_to_write_with_same_name_raises_for_parquet(
         self,
-        datastore: type[ZipBasedPixelFileWithParquet] | type[ZipBasedPixelFileWithCSV],
         setup_basic_pixel_dataset: tuple[
             PixelDataset, DataFrame, AnnData, dict[str, int], DataFrame, DataFrame
         ],
@@ -92,34 +92,86 @@ class TestPixelDataStore:
         file_target = tmp_path / "dataset.pxl"
         dataset.save(
             str(file_target),
-            file_format=(
-                "csv" if datastore.EDGELIST_KEY.endswith(".csv.gz") else "parquet"
-            ),
+            file_format="parquet",
         )
 
-        datastore_inst = datastore(file_target)
+        datastore_inst = ZipBasedPixelFileWithParquet(file_target)
 
-        with pytest.warns(UserWarning, match="Duplicate name: "):
+        with pytest.raises(CannotOverwriteError):
             # Reading something / write something / read something
             adata = datastore_inst.read_anndata()
             datastore_inst.write_anndata(adata)
             _ = datastore_inst.read_anndata()
 
+        with pytest.raises(CannotOverwriteError):
             edgelists = datastore_inst.read_edgelist()
             datastore_inst.write_edgelist(edgelists)
             _ = datastore_inst.read_edgelist()
 
+        with pytest.raises(CannotOverwriteError):
             metadata = datastore_inst.read_metadata()
             datastore_inst.write_metadata(metadata)
             _ = datastore_inst.read_metadata()
 
+        with pytest.raises(CannotOverwriteError):
             polarization = datastore_inst.read_polarization()
             datastore_inst.write_polarization(polarization)
             _ = datastore_inst.read_polarization()
 
+        with pytest.raises(CannotOverwriteError):
             colocalization = datastore_inst.read_colocalization()
             datastore_inst.write_colocalization(colocalization)
             _ = datastore_inst.read_colocalization()
+
+        with pytest.raises(CannotOverwriteError):
+            layouts = datastore_inst.read_precomputed_layouts()
+            datastore_inst.write_precomputed_layouts(layouts)
+            _ = datastore_inst.read_precomputed_layouts()
+
+    def test_pixelfile_datastore_trying_to_write_with_same_name_raises_for_csv(
+        self,
+        setup_basic_pixel_dataset: tuple[
+            PixelDataset, DataFrame, AnnData, dict[str, int], DataFrame, DataFrame
+        ],
+        tmp_path: Path,
+    ):
+        dataset, *_ = setup_basic_pixel_dataset
+        file_target = tmp_path / "dataset.pxl"
+        dataset.save(
+            str(file_target),
+            file_format="csv",
+        )
+
+        datastore_inst = ZipBasedPixelFileWithCSV(file_target)
+
+        with pytest.raises(CannotOverwriteError):
+            # Reading something / write something / read something
+            adata = datastore_inst.read_anndata()
+            datastore_inst.write_anndata(adata)
+            _ = datastore_inst.read_anndata()
+
+        with pytest.raises(CannotOverwriteError):
+            edgelists = datastore_inst.read_edgelist()
+            datastore_inst.write_edgelist(edgelists)
+            _ = datastore_inst.read_edgelist()
+
+        with pytest.raises(CannotOverwriteError):
+            metadata = datastore_inst.read_metadata()
+            datastore_inst.write_metadata(metadata)
+            _ = datastore_inst.read_metadata()
+
+        with pytest.raises(CannotOverwriteError):
+            polarization = datastore_inst.read_polarization()
+            datastore_inst.write_polarization(polarization)
+            _ = datastore_inst.read_polarization()
+
+        with pytest.raises(CannotOverwriteError):
+            colocalization = datastore_inst.read_colocalization()
+            datastore_inst.write_colocalization(colocalization)
+            _ = datastore_inst.read_colocalization()
+
+        # csv files do not support reading/writing layouts
+        # since they require hive-style parquet files
 
     def test_read_metadata(self, pixel_dataset_file: Path):
         datastore = PixelDataStore.guess_datastore_from_path(pixel_dataset_file)
@@ -172,6 +224,52 @@ class TestZipBasedPixelFile:
             "component=PXLCMP0000002",
             "component=PXLCMP0000003",
             "component=PXLCMP0000004",
+        }
+
+    def test_pixelfile_datastore_can_read_layouts(
+        self, tmp_path: Path, layout_df: pd.DataFrame
+    ):
+        file_target = tmp_path / "dataset.pxl"
+
+        precomputed_layout = PreComputedLayouts(pl.DataFrame(layout_df).lazy())
+        with ZipBasedPixelFileWithParquet(file_target) as datastore:
+            datastore.write_precomputed_layouts(precomputed_layout)
+
+        with ZipBasedPixelFileWithParquet(file_target) as datastore:
+            result = datastore.read_precomputed_layouts()
+            assert not result.is_empty
+
+    def test_pixelfile_datastore_can_write_layouts(
+        self,
+        tmp_path: Path,
+        layout_df: pd.DataFrame,
+    ):
+        file_target = tmp_path / "dataset.pxl"
+        precomputed_layout = PreComputedLayouts(pl.DataFrame(layout_df).lazy())
+        with ZipBasedPixelFileWithParquet(file_target) as datastore:
+            datastore.write_precomputed_layouts(layouts=precomputed_layout)
+
+        assert set(ZipFile(file_target).namelist()) == {
+            "layouts.parquet/graph_projection=a-node/layout=fr/component=PXLCMP0000001/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=fr/component=PXLCMP0000000/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=fr/component=PXLCMP0000003/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=fr/component=PXLCMP0000002/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=pmds/component=PXLCMP0000000/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=fr/component=PXLCMP0000000/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=pmds/component=PXLCMP0000004/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=pmds/component=PXLCMP0000004/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=fr/component=PXLCMP0000001/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=pmds/component=PXLCMP0000002/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=pmds/component=PXLCMP0000001/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=pmds/component=PXLCMP0000002/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=fr/component=PXLCMP0000004/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=fr/component=PXLCMP0000002/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=fr/component=PXLCMP0000004/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=pmds/component=PXLCMP0000003/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=pmds/component=PXLCMP0000000/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=pmds/component=PXLCMP0000001/part-0.parquet",
+            "layouts.parquet/graph_projection=a-node/layout=pmds/component=PXLCMP0000003/part-0.parquet",
+            "layouts.parquet/graph_projection=bipartite/layout=fr/component=PXLCMP0000003/part-0.parquet",
         }
 
     def test_pixelfile_datastore_can_write_with_partitioning_with_multiple_partitions(
