+++ tests/pixeldataset/test_utils.py
@@ -1,0 +1,172 @@
+"""Tests for pixeldataset.utils module.
+
+Copyright (c) 2022 Pixelgen Technologies AB.
+"""
+
+# pylint: disable=redefined-outer-name
+
+import logging
+import random
+from pathlib import Path
+from tempfile import TemporaryDirectory
+
+import numpy as np
+import pandas as pd
+from anndata import AnnData
+from numpy.testing import assert_array_equal
+from pandas.testing import assert_frame_equal
+from pixelator.config import AntibodyPanel
+from pixelator.graph import write_recovered_components
+
+from pixelator.pixeldataset.utils import (
+    antibody_metrics,
+    component_antibody_counts,
+    edgelist_to_anndata,
+    read_anndata,
+    write_anndata,
+)
+from pixelator.statistics import (
+    clr_transformation,
+    log1p_transformation,
+)
+
+random.seed(42)
+np.random.seed(42)
+
+
+def test_write_recovered_components(tmp_path: Path):
+    """test_write_recovered_components."""
+    file_target = tmp_path / "components_recovered.csv"
+    write_recovered_components(
+        {"PXLCMP0000000": ["PXLCMP0000000", "PXLCMP0000001"]},
+        filename=file_target,
+    )
+
+    result = pd.read_csv(file_target)
+    assert list(result.columns) == ["cell_id", "recovered_from"]
+    assert_frame_equal(
+        result,
+        pd.DataFrame(
+            {
+                "cell_id": ["PXLCMP0000000", "PXLCMP0000001"],
+                "recovered_from": ["PXLCMP0000000", "PXLCMP0000000"],
+            }
+        ),
+    )
+
+
+def test_antibody_metrics(full_graph_edgelist: pd.DataFrame):
+    """test_antibody_metrics."""
+    assert_frame_equal(
+        antibody_metrics(edgelist=full_graph_edgelist),
+        pd.DataFrame(
+            data={
+                "antibody_count": [1250, 1250],
+                "components": [1, 1],
+                "antibody_pct": [0.5, 0.5],
+            },
+            index=pd.CategoricalIndex(
+                ["A", "B"],
+                name="marker",
+            ),
+        ),
+    )
+
+
+def test_antibody_counts(full_graph_edgelist: pd.DataFrame):
+    """test_antibody_counts."""
+    counts = component_antibody_counts(edgelist=full_graph_edgelist)
+    assert_array_equal(
+        counts.to_numpy(),
+        np.array([[1250, 1250]]),
+    )
+    assert sorted(counts.columns) == sorted(["A", "B"])
+
+
+def test_adata_creation(edgelist: pd.DataFrame, panel: AntibodyPanel):
+    """test_adata_creation."""
+    adata = edgelist_to_anndata(edgelist=edgelist, panel=panel)
+    assert adata.n_vars == panel.size
+    assert adata.n_obs == edgelist["component"].nunique()
+    assert sorted(adata.var) == sorted(
+        [
+            "antibody_count",
+            "components",
+            "antibody_pct",
+            "control",
+            "nuclear",
+        ]
+    )
+    assert sorted(adata.obs) == sorted(
+        [
+            "vertices",
+            "molecules",
+            "antibodies",
+            "upia",
+            "upib",
+            "reads",
+            "mean_reads_per_molecule",
+            "median_reads_per_molecule",
+            "mean_upia_degree",
+            "median_upia_degree",
+            "mean_umi_per_upia",
+            "median_umi_per_upia",
+            "upia_per_upib",
+        ]
+    )
+    assert "clr" in adata.obsm
+    assert "log1p" in adata.obsm
+
+
+def test_read_write_anndata(adata: AnnData):
+    """test_read_write_anndata."""
+    with TemporaryDirectory() as tmp_dir:
+        output_path = Path(tmp_dir) / "example.h5ad"
+        write_anndata(adata, output_path)
+        assert output_path.is_file()
+        adata2 = read_anndata(str(output_path))
+        assert_frame_equal(adata.to_df(), adata2.to_df())
+        assert_frame_equal(adata.obs, adata2.obs)
+        assert_frame_equal(adata.var, adata2.var)
+        assert_array_equal(
+            adata.obsm["clr"],
+            adata2.obsm["clr"],
+        )
+        assert_array_equal(
+            adata.obsm["log1p"],
+            adata2.obsm["log1p"],
+        )
+
+
+def test_edgelist_to_anndata_missing_markers(
+    panel: AntibodyPanel, edgelist: pd.DataFrame, caplog
+):
+    """test_edgelist_to_anndata_missing_markers."""
+    with caplog.at_level(logging.WARN):
+        edgelist_to_anndata(edgelist, panel)
+
+    assert "The given 'panel' is missing markers" in caplog.text
+
+
+def test_edgelist_to_anndata(
+    adata: AnnData, panel: AntibodyPanel, edgelist: pd.DataFrame
+):
+    """test_edgelist_to_anndata."""
+    antibodies = panel.markers
+    counts_df = component_antibody_counts(edgelist=edgelist)
+    counts_df = counts_df.reindex(columns=antibodies, fill_value=0)
+    assert_array_equal(adata.X, counts_df.to_numpy())
+
+    counts_df_clr = clr_transformation(counts_df, axis=1)
+    assert_array_equal(
+        adata.obsm["clr"],
+        counts_df_clr.to_numpy(),
+    )
+
+    counts_df_log1p = log1p_transformation(counts_df)
+    assert_array_equal(
+        adata.obsm["log1p"],
+        counts_df_log1p.to_numpy(),
+    )
+
+    assert set(adata.obs_names) == set(edgelist["component"].unique())
